requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
infinity==1.5
intervals==0.9.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.18.6
pillow==11.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
validators==0.34.0
Werkzeug==3.1.3
WTForms==3.2.1
WTForms-Alchemy==0.19.0
WTForms-Components==0.11.0
